pyproject.toml
[tool.poetry]
name = "payment-wallet"
version = "0.1.0"
description = ""
authors = ["Stephen Mue <8377886+stephen2m@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytz = "2022.6"
Django = "4.1.3"
django-configurations = "2.4"
gunicorn = "20.1.0"
psycopg2-binary = "2.9.5"
dj-database-url = "1.2.0"
django-model-utils = "4.3.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-money = "^3.0.0"
django-searchable-encrypted-fields = "^0.2.1"
requests = "^2.28.1"
gql = {version = "^3.4.0", extras = ["requests"]}
redis = "^4.4.0"

[tool.poetry.dev-dependencies]
ipdb = "0.13.11"
ipython = "8.7.0"
flake8 = "6.0.0"
mock = "4.0.3"
factory-boy = "3.2.1"
django-nose = "1.4.7"
coverage = "6.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
